services/collector-service/requirements.txt
fastapi==0.135.1
starlette==0.49.3
uvicorn==0.34.0
httpx==0.28.1
beautifulsoup4==4.12.3
selenium==4.27.1
kafka-python==2.0.2
six==1.17.0
lz4==4.3.3
pymongo==4.11.1
prometheus-fastapi-instrumentator==7.0.0
opentelemetry-api==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-instrumentation-fastapi==0.51b0
